+++ Apps/Pointing_Task/App2/App.xaml
@@ -1,7 +1,8 @@
 ﻿<Application
-    x:Class="GazeApp.App"
+    x:Class="Pointing_Task.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:GazeApp">
+    xmlns:local="using:Pointing_Task"
+    RequestedTheme="Light">
 
 </Application>
