+++ Apps/Pointing_Task/App2/App.xaml.cs
@@ -15,7 +15,7 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-namespace GazeApp
+namespace Pointing_Task
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
