+++ Apps/Pointing_Task/App2/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
