+++ Apps/Pointing_Task/App2/MainPage.xaml
@@ -1,20 +1,19 @@
 ﻿<Page
-    x:Class="GazeApp.MainPage"
+    x:Class="Pointing_Task.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Pointing_Task"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:gaze="using:Microsoft.Toolkit.Uwp.Input.GazeInteraction"
-    xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities    "
     gaze:GazeInput.Interaction="Enabled"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+    mc:Ignorable="d">
 
-    <Grid>
-        <Grid.RowDefinitions>
+    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <!--<Grid.RowDefinitions>
             <RowDefinition Height="619*"/>
             <RowDefinition Height="381*"/>
-        </Grid.RowDefinitions>
+        </Grid.RowDefinitions>-->
         <Button x:Name="StartButton" Content="Click Me" HorizontalAlignment="Left" Margin="673,565,0,0" VerticalAlignment="Top" RenderTransformOrigin="2.712,3.755" Height="96" UseLayoutRounding="False" Width="170" d:LayoutRounding="Auto" Click="StartButton_Click" Grid.RowSpan="2">
             <Button.RenderTransform>
                 <CompositeTransform Rotation="0"/>
