+++ Apps/Pointing_Task/App2/Package.appxmanifest
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Package xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10" xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest" xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10" IgnorableNamespaces="uap mp">
-  <Identity Name="1c96e33c-70ee-429e-b627-5605a04bfc35" Publisher="CN=Kp" Version="1.0.0.0" />
-  <mp:PhoneIdentity PhoneProductId="1c96e33c-70ee-429e-b627-5605a04bfc35" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
+  <Identity Name="eed847b3-b22a-4156-bd1f-08b94b05bb73" Publisher="CN=Sin-PC" Version="1.1.4.0" />
+  <mp:PhoneIdentity PhoneProductId="eed847b3-b22a-4156-bd1f-08b94b05bb73" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
   <Properties>
-    <DisplayName>GazeApp</DisplayName>
-    <PublisherDisplayName>Kp</PublisherDisplayName>
+    <DisplayName>Pointing_Task</DisplayName>
+    <PublisherDisplayName>Sin-PC</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
   <Dependencies>
@@ -14,8 +14,8 @@
     <Resource Language="x-generate" />
   </Resources>
   <Applications>
-    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="GazeApp.App">
-      <uap:VisualElements DisplayName="GazeApp" Square150x150Logo="Assets\Square150x150Logo.png" Square44x44Logo="Assets\Square44x44Logo.png" Description="GazeApp" BackgroundColor="transparent">
+    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="Pointing_Task.App">
+      <uap:VisualElements DisplayName="Pointing" Square150x150Logo="Assets\Square150x150Logo.png" Square44x44Logo="Assets\Square44x44Logo.png" Description="Pointing Task" BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png">
         </uap:DefaultTile>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
@@ -23,7 +23,7 @@
     </Application>
   </Applications>
   <Capabilities>
+    <Capability Name="internetClient" />
     <DeviceCapability Name="gazeInput" />
-    
   </Capabilities>
 </Package>
