+++ Apps/Pointing_Task/App2/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("GazeApp")]
+[assembly: AssemblyTitle("Pointing_Task")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("GazeApp")]
-[assembly: AssemblyCopyright("Copyright ©  2018")]
+[assembly: AssemblyProduct("Pointing_Task")]
+[assembly: AssemblyCopyright("Copyright I3D,IISC©  2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
