+++ Apps/Pointing_Task/App2/Properties/Default.rd.xml
