+++ Apps/Pointing_Task/Pointing_Task.sln
@@ -1,0 +1,43 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26730.8
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pointing_Task", "App2\Pointing_Task.csproj", "{EAA3DD1D-5417-4E3F-B239-0644A4446379}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|ARM = Debug|ARM
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|ARM = Release|ARM
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|ARM.ActiveCfg = Debug|ARM
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|ARM.Build.0 = Debug|ARM
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|ARM.Deploy.0 = Debug|ARM
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|x64.ActiveCfg = Debug|x64
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|x64.Build.0 = Debug|x64
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|x64.Deploy.0 = Debug|x64
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|x86.ActiveCfg = Debug|x86
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|x86.Build.0 = Debug|x86
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Debug|x86.Deploy.0 = Debug|x86
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|ARM.ActiveCfg = Release|ARM
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|ARM.Build.0 = Release|ARM
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|ARM.Deploy.0 = Release|ARM
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|x64.ActiveCfg = Release|x64
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|x64.Build.0 = Release|x64
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|x64.Deploy.0 = Release|x64
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|x86.ActiveCfg = Release|x86
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|x86.Build.0 = Release|x86
+		{EAA3DD1D-5417-4E3F-B239-0644A4446379}.Release|x86.Deploy.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {94D168FE-4978-4D4B-B6F7-7C6D9378AE48}
+	EndGlobalSection
+EndGlobal
